requirements.txt
Django==4.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
whitenoise==6.4.0